20211223 차트 1차
