+++ src/feature/calculator/ChartSuspense.tsx
